src/frontend/requirements.txt
certifi==2022.6.15
cycler==0.11.0
fonttools==4.33.3
influxdb-client==1.30.0
kiwisolver==1.4.3
matplotlib==3.5.2
numpy==1.23.0
packaging==21.3
pandas==1.4.3
Pillow==9.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
Rx==3.2.0
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
urllib3==1.26.9
